[refactor] rename cogs and clean up some comments
