+++ slashbot/cogs/users.py
@@ -11,7 +11,7 @@ from disnake.ext import commands
 from sqlalchemy.orm import Session
 
 from slashbot.config import App
-from slashbot.cog import CustomCog
+from slashbot.custom_cog import CustomCog
 from slashbot.db import connect_to_database_engine
 from slashbot.db import get_user
 from slashbot.db import BadWord
@@ -27,7 +27,7 @@ USER_OPTIONS = [
 ]
 
 
-class Users(CustomCog):
+class UserCommands(CustomCog):
     """Cog for commands used to save user data."""
 
     def __init__(self, bot: commands.InteractionBot) -> None:
