+++ slashbot/custom_cog.py
